+++ lib/domain/diary_post/diary_post_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:pilll/analytics.dart';
 import 'package:pilll/domain/diary/diary_state.codegen.dart';
-import 'package:pilll/domain/diary_post/post_diary_store.dart';
+import 'package:pilll/domain/diary_post/diary_post_store.dart';
 import 'package:pilll/domain/diary_post/diary_post_store_provider_family.dart';
 import 'package:pilll/entity/diary.codegen.dart';
 import 'package:pilll/database/diary.dart';
